pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hybrid-llm-icpc-lite"
version = "0.1.0"
dependencies = [
  "fastapi",
  "uvicorn[standard]",
  "sqlmodel",
  "pydantic",
  "python-dotenv",
  "playwright",
  "httpx",
  "tenacity",
  "orjson",
  "loguru",
  "cloudscraper",
  "beautifulsoup4",
  "lxml",
]
requires-python = ">=3.8"

[tool.setuptools.packages.find]
include = ["core*", "apps*"]

[tool.playwright]
# Run once:  playwright install chromium
